Added authorization to task controller

+++ app/Http/Controllers/GoalTasksController.php
@@ -25,24 +25,63 @@ class GoalTasksController extends Controller
 
     }
 
-    public function showAll($goal_id)
+    public function showAllTasks($goal_id)
     {
-        $user = Auth::user();
+            $goalExist = Goal::where('id', $goal_id)->exists();
+            $goal = Goal::findOrFail($goal_id);
 
-        $tasks =  Task::where('goal_id', $goal_id)->get();
+            if (Auth::user()->id == $goal->user_id){
 
-        return response()->json(['data' => ['success' => true, 'tasks' => $tasks], 200]);
+                if ($goalExist){
+
+                    $tasks = Task::where('goal_id', $goal_id)->get();
+
+                    if ($tasks){
+
+                        return response()->json(['data' => ['success' => true, 'tasks' => $tasks], 200]);
+                    
+                    } else{
+                        return response()->json('Task Does not Exist', 401);
+                    }
+
+                } else{
+                    return response()->json('Goal Does not Exist', 401);
+                }
+
+            } else{
+                return response()->json('Unauthorized', 401);
+            }
 
     }
 
-    public function showOne($goal_id, $task_id)
+    public function showOneTask($goal_id, $task_id)
     {
         $user = Auth::user();
+        $goal = Goal::findOrFail($goal_id);
+        $goalExist = Goal::where('id', $goal_id)->exists();
+        $taskExist = Task::where('id', $task_id)->exists();
 
-        $task =  Task::where('goal_id', $goal_id)->where('id', $task_id)->first();
+        if ($user->id == $goal->user_id){
 
-        return response()->json(['data' => ['success' => true, 'task' => $task], 200]);
+            if ($goalExist && $taskExist){
+                    
+                $task = Task::where('id', $task_id)->where('goal_id', $goal_id)->first();
+                    
+                    if ($task){
+                    
+                        return response()->json(['data' => ['success' => true, 'task' => $task], 200]);
 
+                    } else {
+                            return response()->json('Unauthorized!', 400);
+                    }
+                
+                } else {
+                        return response()->json('Goal or Task Does not exist!', 400);
+                }
+
+            } else {
+                return response()->json('Unauthorized', 400);
+            }
     }
 
     public function store(Request $request, $goal_id)
@@ -84,27 +123,33 @@ class GoalTasksController extends Controller
 
     }
 
-    public function update(Request $request, $task_id)
+    public function update(Request $request, $task_id, $goal_id)
     {
         
         $user = Auth::user();
+        $goal = Goal::findOrFail($goal_id);
 
         try {
+
+            if($user->id ==  $goal->user_id){
+
+                $task = Task::findOrFail($task_id);
+
+                $task->title = $request->input('title');
+                $task->description = $request->input('description');
+                $task->completed = $request->input('completed');
+                $task->start = $request->input('start');
+                $task->finish = $request->input('finish');
+
+                $task->save();
+
+                $res['status'] = True;
+                $res['data'] = "$task->title Updated Successfully";
+                $res['task'] = $task;
+                return response()->json($res, 200);
+
+            } else return response()->json('Unauthorized!!!');
             
-            $task = Task::findOrFail($task_id);
-
-            $task->title = $request->input('title');
-            $task->description = $request->input('description');
-            $task->completed = $request->input('completed');
-            $task->start = $request->input('start');
-            $task->finish = $request->input('finish');
-
-            $task->save();
-
-            $res['status'] = True;
-            $res['data'] = "$task->title Updated Successfully";
-            $res['task'] = $task;
-            return response()->json($res, 200);
 
         } catch (\Exception $e) {
             
@@ -114,16 +159,23 @@ class GoalTasksController extends Controller
         
     }
 
-    public function destroy(Request $request, $task_id)
+    public function destroy(Request $request, $task_id, $goal_id)
     {
         $user = Auth::user();
+        $goal = Goal::findOrFail($goal_id);
 
-        $task = Task::findOrFail($task_id);
+        if($user->id == $goal->user_id){
 
-        $task->delete();
+            $task = Task::findOrFail($task_id);
 
-        return response('Deleted Successfully', 200);
+            $task->delete();
 
+            return response('Deleted Successfully', 200);
+
+        } else {
+            return response()->json('Unauthorized!!!');
+        }
+    
     }
  
 }
